Add exception handling and getById

+++ backend/src/main/kotlin/com/treska/yummier/service/RecipeService.kt
@@ -1,6 +1,7 @@
 package com.treska.yummier.service
 
 import com.treska.yummier.dto.RecipeFilter
+import com.treska.yummier.exception.RecipeNotFoundException
 import com.treska.yummier.extension.normalize
 import com.treska.yummier.model.Category
 import com.treska.yummier.model.Difficulty
@@ -15,6 +16,13 @@ import org.springframework.stereotype.Service
 class RecipeService(private val recipeRepository: RecipeRepository, private val ingredientService: IngredientService) {
     fun get(filter: RecipeFilter, pageable: Pageable): Page<Recipe> {
         return recipeRepository.findAll(SpecificationBuilder.withFilters(filter), pageable)
+    }
+
+    fun getById(id: Long): Recipe {
+        if (!recipeRepository.existsById(id)) {
+            throw RecipeNotFoundException("Recipe with id=[$id] does not exist.")
+        }
+        return recipeRepository.findById(id).get()
     }
 
     fun create(
@@ -44,6 +52,9 @@ class RecipeService(private val recipeRepository: RecipeRepository, private val 
     }
 
     fun delete(id: Long) {
+        if(!recipeRepository.existsById(id)) {
+            throw RecipeNotFoundException("Recipe with id=[$id] does not exist.")
+        }
         recipeRepository.deleteById(id)
     }
 }
